Add Pundit policy

+++ app/controllers/admin/law_proposals_controller.rb
@@ -1,17 +1,25 @@
 # frozen_string_literal: true
 
 module Admin
-  class LawProposalsController < ApplicationController
+  class LawProposalsController < ::AdminController
     expose(:law_proposals) { LawProposal.all }
     expose(:law_proposal)
 
-    def index; end
+    def index
+      policy_scope(LawProposal)
+    end
 
-    def show; end
+    def show
+      authorize law_proposal
+    end
 
-    def new; end
+    def new
+      authorize law_proposal
+    end
 
     def create
+      authorize law_proposal
+
       if law_proposal.save
         redirect_to admin_law_proposals_path
       else
@@ -19,9 +27,13 @@ module Admin
       end
     end
 
-    def edit; end
+    def edit
+      authorize law_proposal
+    end
 
     def update
+      authorize law_proposal
+
       if law_proposal.update(law_proposal_params)
         redirect_to admin_law_proposals_path
       else
@@ -30,6 +42,8 @@ module Admin
     end
 
     def delete
+      authorize law_proposal
+
       law_proposal.destroy
       redirect_to :index, status: :ok
     end
